Service layer

+++ src/main/java/com/dogitalcode/reactive/repository/StudentRepository.java
@@ -4,7 +4,7 @@ import com.dogitalcode.reactive.model.Student;
 import org.springframework.data.repository.reactive.ReactiveCrudRepository;
 import reactor.core.publisher.Flux;
 
-public interface StudentRepository extends ReactiveCrudRepository<Student, Integer> {
+public interface StudentRepository extends ReactiveCrudRepository<Student, Long> {
 
     Flux<Student> findAllByFirstnameContainingIgnoreCase(String firstName);
 }
